requirements.txt
Django
djangorestframework
python-dotenv
groq
psycopg2-binary
django-redis
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
